inference/source/requirements.txt
torch==2.0.1
transformers==4.36.2
peft==0.11.1
bitsandbytes==0.43.1
boto3==1.34.131
numpy==1.26.4
torchvision==0.15.2
pillow==10.4.0